% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Locations of traffic accidents in the USA}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state identification}

\item{year}{reference year}
}
\value{
Dataset in Tibble form.
}
\description{
Maps accident locations in each state by year
}
\examples{
\dontrun{
fars_map_state(12,2013)
}
}
