% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read a .csv file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{File location or name if in working directory}
}
\value{
Dataset in Tibble
}
\description{
Reads a .csv file and turns it into a tibble
}
\examples{
\dontrun{
fars_read("R/accident_2013.csv.bz2")
fars_read("./accident_2013.csv.bz2")
}
}
