% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extracts the month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{reference year}
}
\value{
Dataset in Tibble form.
}
\description{
Extracts the month and year from a data set
}
\examples{
\dontrun{
fars_read_years(2015)
}
}
