% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Extract the number of vehicle collision injuries per month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{reference year}
}
\value{
Dataset in Tibble form.
}
\description{
Extract the number of vehicle collision injuries per month with reference year
}
\examples{
\dontrun{
fars_summarize_years(2013)
}
}
