% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Prints the name of a file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year that complements the character string}
}
\value{
Chraracter with combination
}
\description{
Prints the name of a file from the combination of a character string and value of interest
}
\examples{
\dontrun{
make_filename(2013)
}
}
